Added a new "Relation" class in a relation.py file. Its job is to track diplomatic relations between nations.
The first proper diplomatic relation added is of course a war. Currently it's buggy as hell. Also added rudimentary possibilities for declaring war and invading. Needs to be fixed, but I need to lie in bed before my document check to gain some energy, so will fully implement it later.